simplify DialogHeader component
by combining title/description into single header component with description as
a prop

also makes structure same as AlertDialog component now
