+++ .changeset/empty-lizards-fix.md
@@ -1,0 +1,6 @@
+---
+"@optiaxiom/react": minor
+"@optiaxiom/web-components": minor
+---
+
+simplify DialogHeader component
