+++ apps/docs/demos/dialog/description-usage/App.tsx
@@ -2,10 +2,8 @@ import {
   Dialog,
   DialogClose,
   DialogContent,
-  DialogDescription,
   DialogFooter,
   DialogHeader,
-  DialogTitle,
   DialogTrigger,
 } from "@optiaxiom/react";
 
@@ -15,11 +13,8 @@ export function App() {
       <DialogTrigger>Open Dialog</DialogTrigger>
 
       <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Modal Title</DialogTitle>
-          <DialogDescription>
-            This is additional description of the modal
-          </DialogDescription>
+        <DialogHeader description="This is additional description of the modal">
+          Modal Title
         </DialogHeader>
         <DialogFooter>
           <DialogClose>Close</DialogClose>
