--- packages/react/src/dialog-actions/DialogActions.css.ts
@@ -1,13 +1,0 @@
-import { recipe, style } from "../vanilla-extract";
-
-export const actions = recipe({
-  base: [
-    {
-      flexDirection: "row",
-      gap: "8",
-    },
-    style({
-      order: "2",
-    }),
-  ],
-});
