--- packages/react/src/dialog-actions/index.ts
@@ -1,1 +1,0 @@
-export * from "./DialogActions";
