--- packages/react/src/dialog-description/index.ts
@@ -1,1 +1,0 @@
-export * from "./DialogDescription";
