+++ packages/react/src/dialog-header/DialogHeader.css.ts
@@ -7,3 +7,40 @@ export const close = recipe({
     }),
   ],
 });
+
+export const title = recipe({
+  base: [
+    {
+      flex: "1",
+      fontWeight: "500",
+    },
+    style({
+      order: "1",
+    }),
+  ],
+});
+
+export const actions = recipe({
+  base: [
+    {
+      flexDirection: "row",
+      gap: "8",
+    },
+    style({
+      order: "2",
+    }),
+  ],
+});
+
+export const description = recipe({
+  base: [
+    {
+      color: "fg.secondary",
+      fontWeight: "400",
+      w: "full",
+    },
+    style({
+      order: "4",
+    }),
+  ],
+});
