+++ packages/react/src/dialog-header/DialogHeader.tsx
@@ -1,15 +1,25 @@
 import * as RadixDialog from "@radix-ui/react-dialog";
-import { forwardRef } from "react";
+import { VisuallyHidden } from "@radix-ui/react-visually-hidden";
+import { forwardRef, type ReactNode } from "react";
 
 import { Box, type BoxProps } from "../box";
 import { Button } from "../button";
+import { Flex } from "../flex";
+import { Heading } from "../heading";
 import { IconX } from "../icons/IconX";
+import { Text } from "../text";
 import * as styles from "./DialogHeader.css";
 
-type DialogHeaderProps = BoxProps;
+type DialogHeaderProps = BoxProps<
+  "div",
+  {
+    addonAfter?: ReactNode;
+    description?: ReactNode;
+  }
+>;
 
 export const DialogHeader = forwardRef<HTMLHeadingElement, DialogHeaderProps>(
-  ({ children, ...props }, ref) => {
+  ({ addonAfter, children, description, ...props }, ref) => {
     return (
       <Box
         display="flex"
@@ -30,7 +40,21 @@ export const DialogHeader = forwardRef<HTMLHeadingElement, DialogHeaderProps>(
           <RadixDialog.Close />
         </Button>
 
-        {children}
+        <Heading asChild level="3" {...styles.title()}>
+          <RadixDialog.Title>{children}</RadixDialog.Title>
+        </Heading>
+
+        {addonAfter && <Flex {...styles.actions()}>{addonAfter}</Flex>}
+
+        {description ? (
+          <Text asChild {...styles.description()}>
+            <RadixDialog.Description>{description}</RadixDialog.Description>
+          </Text>
+        ) : (
+          <VisuallyHidden>
+            <RadixDialog.Description />
+          </VisuallyHidden>
+        )}
       </Box>
     );
   },
