--- packages/react/src/dialog-title/DialogTitle.css.ts
@@ -1,13 +1,0 @@
-import { recipe, style } from "../vanilla-extract";
-
-export const title = recipe({
-  base: [
-    {
-      flex: "1",
-      fontWeight: "500",
-    },
-    style({
-      order: "1",
-    }),
-  ],
-});
