--- packages/react/src/dialog-title/index.ts
@@ -1,1 +1,0 @@
-export * from "./DialogTitle";
