+++ packages/react/src/index.ts
@@ -18,15 +18,12 @@ export * from "./button-group";
 export * from "./checkbox";
 export * from "./code";
 export * from "./dialog";
-export * from "./dialog-actions";
 export * from "./dialog-body";
 export * from "./dialog-close";
 export * from "./dialog-content";
-export * from "./dialog-description";
 export * from "./dialog-footer";
 export * from "./dialog-form";
 export * from "./dialog-header";
-export * from "./dialog-title";
 export * from "./dialog-trigger";
 export * from "./disclosure";
 export * from "./disclosure-content";
