+++ packages/web-components/package.json
@@ -35,15 +35,12 @@
     "./Checkbox": "./dist/components/Checkbox.js",
     "./Code": "./dist/components/Code.js",
     "./Dialog": "./dist/components/Dialog.js",
-    "./DialogActions": "./dist/components/DialogActions.js",
     "./DialogBody": "./dist/components/DialogBody.js",
     "./DialogClose": "./dist/components/DialogClose.js",
     "./DialogContent": "./dist/components/DialogContent.js",
-    "./DialogDescription": "./dist/components/DialogDescription.js",
     "./DialogFooter": "./dist/components/DialogFooter.js",
     "./DialogForm": "./dist/components/DialogForm.js",
     "./DialogHeader": "./dist/components/DialogHeader.js",
-    "./DialogTitle": "./dist/components/DialogTitle.js",
     "./DialogTrigger": "./dist/components/DialogTrigger.js",
     "./Disclosure": "./dist/components/Disclosure.js",
     "./DisclosureContent": "./dist/components/DisclosureContent.js",
